修改 类型转换错误

+++ zeus-webapp/src/main/java/com/zmops/iot/web/product/service/work/SaveProdAttrWorker.java
@@ -32,7 +32,7 @@ public class SaveProdAttrWorker implements IWorker<ProductAttr, Boolean> {
 
         String prodId = productAttr.getProductId();
 
-        List<String> deviceIds = new QDevice().select(QDevice.Alias.deviceId).productId.eq(Integer.parseInt(prodId)).findSingleAttributeList();
+        List<String> deviceIds = new QDevice().select(QDevice.Alias.deviceId).productId.eq(Long.parseLong(prodId)).findSingleAttributeList();
 
         List<ProductAttribute> productAttributeList = new ArrayList<>();
